refactor(i18n): move consts
