+++ libs/common/src/lib/i18n/index.ts
@@ -1,0 +1,3 @@
+export * from './i18n.module'
+export * from './i18n.constants'
+export * from './i18n.interceptor'
